Time: 0 ms (100%), Space: 8.1 MB (42.94%) - LeetHub

+++ 1344-angle-between-hands-of-a-clock/1344-angle-between-hands-of-a-clock.cpp
@@ -1,14 +1,14 @@
 class Solution {
 public:
     double angleClock(int hour, int minutes) {
-       double  mindeg=6.00*minutes;
+       float  mindeg=6.00*minutes;
      
-       double hrdg=30.00*hour;
+       float hrdg=30.00*hour;
         if (hour==12) hrdg=0;
 
-        double extramovehr=0.5*minutes;
+        float extramovehr=0.5*minutes;
 
-       double angle= abs(hrdg+extramovehr-mindeg);
+       float angle= abs(hrdg+extramovehr-mindeg);
        return min(angle, 360-angle);
     }
 };
